More vim fixes and move some more things out of app.rs
