+++ assets/keymaps/vim.json
@@ -315,7 +315,7 @@
     {
         "context": "Editor && VimWaiting",
         "bindings": {
-            "*": "gpui::KeyPressed",
+            // "*": "gpui::KeyPressed",
             "escape": "editor::Cancel"
         }
     }
