+++ crates/sqlez/Cargo.toml
@@ -16,6 +16,3 @@ lazy_static = "1.4"
 parking_lot = "0.11.1"
 futures = "0.3"
 uuid = { version = "1.1.2", features = ["v4"] }
-
-[dev-dependencies]
-sqlez_macros = { path = "../sqlez_macros"}
